Make the UI prettier
The UI was really ugly and hard to read.  This was on account of an ill-
suited font and the transparent panels made it hard to read.  As well,
the license in the settings menu provided an enormous wall of text that
wasn't doing it any favors.  All of these things were adjusted and now
it's pretty.

The font chosen was Computer Modern Bright, which was designed for TeX
by Donald Knuth.  It's available under the SIL Open Font License.  I've
relicensed it here under GPLv3+ and have added a font exception.  As
well, I've modified the font to only include ASCII characters (U+0000
to U+007f) as a space-saving measure.