Update to Godot 4
There were two steps to this--the automatic update and the manual
update.  The automatic update was Godot's automatic system for migrating
Godot 3 projects, and then I had to go in and fix things that had gone
wrong.

The principal things I needed to fix were that the way TileMaps work in
Godot 4 is very different.  Tile coordinates are Vector2i instead of
Vector2 and this led to some logic issues.  There were also API changes
in TileMap that I had to deal with.

TileSets have also changed, so I needed to update that.  With the
automatic migration, the tiles were far too small for what I expected.
It had the tile size as 16x16, but the tiles had been 100x100.  But it
was easy to fix.

The diff is quite large because the migration tool appears to have
migrated my spaces to tabs.  A bold move to be sure.